content/html/content/test/test_bug332246.html

<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=332246
-->
<head>
  <title>Test for Bug 332246 - scrollIntoView(false) doesn't work correctly for inline elements that wrap at multiple lines</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=332246">Mozilla Bug 332246</a>
<p id="display"></p>
<div id="content">

<div id="a1" style="height: 100px; width: 100px; overflow: hidden;">
<div style="height: 100px"></div>
<a id="a2" href="#">Top<br><br><br><br><br><br><br><br><br><br><br><br>Bottom</a>
<div style="height: 100px"></div>
</div>

<div id="b1" style="height: 100px; width: 100px; overflow: hidden;">
<div style="height: 100px"></div>
<div id="b2" style="border: 10px solid black; height: 200px;"></div>
<div style="height: 100px"></div>
</div>

<br>

<div id="c1" style="height: 100px; width: 100px; overflow: hidden; position: relative;">
<div id="c2" style="border: 10px solid black; height: 200px; width: 50px; position: absolute; top: 100px;"></div>
<div style="height: 100px"></div>
</div>

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 332246 **/

var a1 = document.getElementById('a1');
var a2 = document.getElementById('a2');
a2.scrollIntoView(true);
//bug 401904 for windows, bug 411297 for the Mac
if (navigator.platform == "Win32" || navigator.platform.indexOf("Mac") == 0)
  todo(a1.scrollTop == 100, "Wrong scrollTop value!");
else
  ok(a1.scrollTop == 100, "Wrong scrollTop value!");
a2.scrollIntoView(false);
if (navigator.platform.indexOf("Mac") == 0) //Mac bug 411297 again
  todo((a1.scrollHeight - a1.offsetHeight - a1.scrollTop) == 100, "Wrong (scrollHeight - offsetHeight - scrollTop) value!");
else
  ok((a1.scrollHeight - a1.offsetHeight - a1.scrollTop) == 100, "Wrong (scrollHeight - offsetHeight - scrollTop) value!");

var b1 = document.getElementById('b1');
var b2 = document.getElementById('b2');
b2.scrollIntoView(true);
  ok(b1.scrollTop == 100, "Wrong scrollTop value!");
b2.scrollIntoView(false);
  ok((b1.scrollHeight - b1.offsetHeight - b1.scrollTop) == 100, "Wrong (scrollHeight - offsetHeight - scrollTop) value!");

var c1 = document.getElementById('c1');
var c2 = document.getElementById('c2');
c2.scrollIntoView(true);
  ok(c1.scrollTop == 100, "Wrong scrollTop value!");
c2.scrollIntoView(false);
  ok((c1.scrollHeight - c1.offsetHeight - c1.scrollTop) == 0, "Wrong (scrollHeight - offsetHeight - scrollTop) value!");

</script>
</pre>
</body>
</html>
